+++ android/anc/src/test/java/org/smartregister/fhircore/anc/ui/family/FamilyRegisterActivityTest.kt
@@ -17,12 +17,9 @@
 package org.smartregister.fhircore.anc.ui.family
 
 import android.app.Activity
-import android.app.Application
-import android.content.Intent
 import android.view.View
 import android.widget.ImageButton
 import android.widget.TextView
-import androidx.test.core.app.ApplicationProvider
 import com.google.android.fhir.sync.Sync
 import io.mockk.every
 import io.mockk.mockk
@@ -32,32 +29,29 @@ import java.time.OffsetDateTime
 import kotlinx.coroutines.flow.flowOf
 import org.junit.After
 import org.junit.Assert.assertEquals
+import org.junit.Assert.assertFalse
 import org.junit.Assert.assertNotNull
 import org.junit.Assert.assertTrue
 import org.junit.Before
 import org.junit.BeforeClass
 import org.junit.Test
 import org.robolectric.Robolectric
-import org.robolectric.Shadows
 import org.robolectric.annotation.Config
 import org.robolectric.fakes.RoboMenuItem
 import org.robolectric.util.ReflectionHelpers
 import org.smartregister.fhircore.anc.R
-import org.smartregister.fhircore.anc.activity.ActivityRobolectricTest
+import org.smartregister.fhircore.anc.activity.BaseRegisterActivityTest
 import org.smartregister.fhircore.anc.data.family.FamilyRepository
 import org.smartregister.fhircore.anc.data.patient.PatientRepository
 import org.smartregister.fhircore.anc.shadow.AncApplicationShadow
 import org.smartregister.fhircore.anc.shadow.FakeKeyStore
 import org.smartregister.fhircore.anc.ui.anccare.register.AncRegisterFragment
-import org.smartregister.fhircore.anc.ui.family.form.FamilyFormConstants
-import org.smartregister.fhircore.anc.ui.family.form.FamilyQuestionnaireActivity
 import org.smartregister.fhircore.anc.ui.family.register.FamilyRegisterActivity
 import org.smartregister.fhircore.anc.ui.family.register.FamilyRegisterFragment
-import org.smartregister.fhircore.engine.ui.questionnaire.QuestionnaireActivity.Companion.QUESTIONNAIRE_ARG_FORM
 import org.smartregister.fhircore.engine.ui.userprofile.UserProfileFragment
 
 @Config(shadows = [AncApplicationShadow::class])
-internal class FamilyRegisterActivityTest : ActivityRobolectricTest() {
+internal class FamilyRegisterActivityTest : BaseRegisterActivityTest() {
 
   private lateinit var familyRegisterActivity: FamilyRegisterActivity
   private lateinit var patientRepository: PatientRepository
@@ -87,24 +81,6 @@ internal class FamilyRegisterActivityTest : ActivityRobolectricTest() {
   @Test
   fun testActivityShouldNotNull() {
     assertNotNull(familyRegisterActivity)
-  }
-
-  @Test
-  fun testRegisterClientShouldStartFamilyQuestionnaireActivity() {
-    ReflectionHelpers.callInstanceMethod<FamilyRegisterActivity>(
-      familyRegisterActivity,
-      "registerClient"
-    )
-
-    val expectedIntent = Intent(familyRegisterActivity, FamilyQuestionnaireActivity::class.java)
-    val actualIntent =
-      Shadows.shadowOf(ApplicationProvider.getApplicationContext<Application>()).nextStartedActivity
-
-    assertEquals(expectedIntent.component, actualIntent.component)
-    assertEquals(
-      FamilyFormConstants.FAMILY_REGISTER_FORM,
-      actualIntent.getStringExtra(QUESTIONNAIRE_ARG_FORM)
-    )
   }
 
   @Test
@@ -151,6 +127,29 @@ internal class FamilyRegisterActivityTest : ActivityRobolectricTest() {
       View.GONE,
       familyRegisterActivity.findViewById<ImageButton>(R.id.filter_register_button).visibility
     )
+  }
+
+  @Test
+  fun testRegistersListShouldReturnFamilyAndAncRegisterItemList() {
+    val listRegisterItem = familyRegisterActivity.registersList()
+    assertEquals(listRegisterItem.size, 2)
+
+    with(listRegisterItem[0]) {
+      assertEquals(FamilyRegisterFragment.TAG, uniqueTag)
+      assertEquals(familyRegisterActivity.getString(R.string.families), title)
+      assertTrue(isSelected)
+    }
+
+    with(listRegisterItem[1]) {
+      assertEquals(AncRegisterFragment.TAG, uniqueTag)
+      assertEquals(familyRegisterActivity.getString(R.string.anc_clients), title)
+      assertFalse(isSelected)
+    }
+  }
+
+  @Test
+  fun testMainFragmentTagShouldReturnFamilyRegisterFragmentTag() {
+    assertEquals(FamilyRegisterFragment.TAG, familyRegisterActivity.mainFragmentTag())
   }
 
   override fun getActivity(): Activity {