+++ android/eir/src/test/java/org/smartregister/fhircore/eir/ui/patient/register/PatientRegisterActivityTest.kt
@@ -17,10 +17,17 @@
 package org.smartregister.fhircore.eir.ui.patient.register
 
 import android.app.Activity
+import android.app.Application
+import android.content.Intent
 import androidx.core.content.ContextCompat
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.Observer
+import androidx.test.core.app.ApplicationProvider
 import io.mockk.every
 import io.mockk.mockk
+import io.mockk.slot
 import io.mockk.spyk
+import io.mockk.verify
 import kotlinx.coroutines.test.runBlockingTest
 import org.junit.Assert
 import org.junit.Before
@@ -37,6 +44,7 @@ import org.smartregister.fhircore.eir.coroutine.CoroutineTestRule
 import org.smartregister.fhircore.eir.shadow.EirApplicationShadow
 import org.smartregister.fhircore.eir.shadow.FakeKeyStore
 import org.smartregister.fhircore.eir.shadow.ShadowNpmPackageProvider
+import org.smartregister.fhircore.eir.ui.patient.details.PatientDetailsActivity
 import org.smartregister.fhircore.engine.ui.register.model.SideMenuOption
 
 @Config(shadows = [EirApplicationShadow::class, ShadowNpmPackageProvider::class])
@@ -93,6 +101,42 @@ class PatientRegisterActivityTest : ActivityRobolectricTest() {
 
     Assert.assertEquals(1, fragments.size)
     Assert.assertTrue(fragments.containsKey(PatientRegisterFragment.TAG))
+  }
+
+  @Test
+  fun testMainFragmentTagShouldReturnPatientRegisterFragmentTag() {
+    Assert.assertEquals(PatientRegisterFragment.TAG, patientRegisterActivity.mainFragmentTag())
+  }
+
+  @Test
+  fun testOnBarcodeResultShouldNavigateToDetailScreenAndRegisterClientScreen() {
+    val activity = spyk(patientRegisterActivity)
+    val data = mockk<MutableLiveData<Result<Boolean>>>()
+    val observer = slot<Observer<Result<Boolean>>>()
+
+    every { activity.isPatientExists(any()) } returns data
+    every { data.observe(any(), capture(observer)) } returns Unit
+    every { activity.navigateToDetails(any()) } returns Unit
+    every { activity.registerClient(any()) } returns Unit
+
+    activity.onBarcodeResult("12345", mockk())
+
+    observer.captured.onChanged(Result.success(true))
+    verify(exactly = 1) { activity.navigateToDetails("12345") }
+
+    observer.captured.onChanged(Result.failure(mockk()))
+    verify(exactly = 1) { activity.registerClient("12345") }
+  }
+
+  @Test
+  fun testNavigateToDetailsShouldNavigateToPatientDetailActivity() {
+    patientRegisterActivity.navigateToDetails("12345")
+
+    val actualComponent = Intent(patientRegisterActivity, PatientDetailsActivity::class.java)
+    val expectedIntent =
+      shadowOf(ApplicationProvider.getApplicationContext<Application>()).nextStartedActivity
+
+    Assert.assertEquals(actualComponent.component, expectedIntent.component)
   }
 
   override fun getActivity(): Activity {