+++ android/engine/src/main/java/org/smartregister/fhircore/engine/ui/questionnaire/QuestionnaireActivity.kt
@@ -27,15 +27,17 @@ import androidx.fragment.app.commit
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.lifecycleScope
 import ca.uhn.fhir.context.FhirContext
+import com.google.android.fhir.FhirEngineProvider.fhirEngine
 import com.google.android.fhir.datacapture.QuestionnaireFragment
 import com.google.android.fhir.datacapture.QuestionnaireFragment.Companion.BUNDLE_KEY_QUESTIONNAIRE
 import com.google.android.fhir.datacapture.QuestionnaireFragment.Companion.BUNDLE_KEY_QUESTIONNAIRE_RESPONSE
 import com.google.android.fhir.datacapture.validation.QuestionnaireResponseValidator
-import com.google.android.fhir.datacapture.validation.ValidationResult
+import com.google.android.fhir.db.ResourceNotFoundException
 import com.google.android.fhir.logicalId
 import org.hl7.fhir.r4.model.Patient
 import org.hl7.fhir.r4.model.Questionnaire
 import org.hl7.fhir.r4.model.QuestionnaireResponse
+import org.hl7.fhir.r4.model.StringType
 import org.smartregister.fhircore.engine.R
 import org.smartregister.fhircore.engine.ui.base.AlertDialogue.showConfirmAlert
 import org.smartregister.fhircore.engine.ui.base.AlertDialogue.showErrorAlert
@@ -48,7 +50,6 @@ import org.smartregister.fhircore.engine.util.extension.createFactory
 import org.smartregister.fhircore.engine.util.extension.find
 import org.smartregister.fhircore.engine.util.extension.showToast
 import timber.log.Timber
-import java.util.stream.Stream
 
 /**
  * Launches Questionnaire with given id. If questionnaire has subjectType = Patient his activity can
@@ -69,6 +70,8 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
 
   protected var clientIdentifier: String? = null
 
+  protected var immunizationId: String? = null
+
   private val parser = FhirContext.forR4().newJsonParser()
 
   override fun onCreate(savedInstanceState: Bundle?) {
@@ -83,14 +86,16 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
     val loadProgress = showProgressAlert(this, R.string.loading)
 
     clientIdentifier = intent.getStringExtra(QUESTIONNAIRE_ARG_PATIENT_KEY)
+    immunizationId = intent.getStringExtra(ADVERSE_EVENT_IMMUNIZATION_ITEM_KEY)
 
     lifecycleScope.launchWhenCreated {
       questionnaireViewModel = createViewModel(application)
 
       val form = intent.getStringExtra(QUESTIONNAIRE_ARG_FORM)!!
       // form is either name of form in asset/form-config or questionnaire-id
-      // load from db or assets
-      questionnaireConfig = kotlin.runCatching { questionnaireViewModel.getQuestionnaireConfig(form) }.getOrElse {
+      // load from assets and get questionnaire or if not found build it from questionnaire
+      questionnaireConfig =
+        kotlin.runCatching { questionnaireViewModel.getQuestionnaireConfig(form) }.getOrElse {
           // load questionnaire from db and build config
           questionnaire = questionnaireViewModel.loadQuestionnaire(form)!!
 
@@ -113,21 +118,33 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
       // Only add the fragment once, when the activity is first created.
       if (savedInstanceState == null) {
         val fragment =
-          QuestionnaireFragment().apply {
+          FhirCoreQuestionnaireFragment().apply {
             val parsedQuestionnaire = parser.encodeResourceToString(questionnaire)
             arguments =
               when {
                 clientIdentifier == null ->
                   bundleOf(Pair(BUNDLE_KEY_QUESTIONNAIRE, parsedQuestionnaire))
                 clientIdentifier != null -> {
-                  val parsedQuestionnaireResponse =
-                    parser.encodeResourceToString(
-                      questionnaireViewModel.generateQuestionnaireResponse(questionnaire!!, intent)
+
+                  try {
+                    fhirEngine.load(Patient::class.java, clientIdentifier!!)
+                    val parsedQuestionnaireResponse =
+                      parser.encodeResourceToString(
+                        questionnaireViewModel.generateQuestionnaireResponse(
+                          questionnaire!!,
+                          intent
+                        )
+                      )
+                    bundleOf(
+                      Pair(BUNDLE_KEY_QUESTIONNAIRE, parser.encodeResourceToString(questionnaire)),
+                      Pair(BUNDLE_KEY_QUESTIONNAIRE_RESPONSE, parsedQuestionnaireResponse)
                     )
-                  bundleOf(
-                    Pair(BUNDLE_KEY_QUESTIONNAIRE, parsedQuestionnaire),
-                    Pair(BUNDLE_KEY_QUESTIONNAIRE_RESPONSE, parsedQuestionnaireResponse)
-                  )
+                  } catch (e: ResourceNotFoundException) {
+                    setBarcode(questionnaire, clientIdentifier!!, true)
+                    bundleOf(
+                      Pair(BUNDLE_KEY_QUESTIONNAIRE, parser.encodeResourceToString(questionnaire))
+                    )
+                  }
                 }
                 else -> bundleOf(Pair(BUNDLE_KEY_QUESTIONNAIRE, parsedQuestionnaire))
               }
@@ -138,6 +155,14 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
     }
 
     findViewById<Button>(R.id.btn_save_client_info).setOnClickListener(this)
+  }
+
+  private fun setBarcode(questionnaire: Questionnaire, code: String, readonly: Boolean) {
+    questionnaire.find(QUESTIONNAIRE_ARG_BARCODE_KEY)?.apply {
+      initial =
+        mutableListOf(Questionnaire.QuestionnaireItemInitialComponent().setValue(StringType(code)))
+      readOnly = readonly
+    }
   }
 
   open fun createViewModel(application: Application) =
@@ -172,13 +197,6 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
     saveProcessingAlertDialog.dismiss()
   }
 
-  override fun onStop() {
-    super.onStop()
-    if (::saveProcessingAlertDialog.isInitialized && saveProcessingAlertDialog != null) {
-      saveProcessingAlertDialog.dismiss()
-    }
-  }
-
   open fun handleQuestionnaireSubmit() {
     saveProcessingAlertDialog = showProgressAlert(this, R.string.saving_registration)
 
@@ -200,24 +218,25 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
     questionnaireViewModel.extractionProgress.observe(
       this,
       { result ->
+        saveProcessingAlertDialog.dismiss()
         if (result) {
           finish()
         } else {
           Timber.e("An error occurred during extraction")
         }
-
-        saveProcessingAlertDialog.dismiss()
       }
     )
   }
 
+  // TODO remove this when SDK bug for validation is fixed
+  // https://github.com/google/android-fhir/issues/912
   fun deepFlat(
     qItems: List<Questionnaire.QuestionnaireItemComponent>,
     questionnaireResponse: QuestionnaireResponse,
     targetQ: MutableList<Questionnaire.QuestionnaireItemComponent>,
     targetQR: MutableList<QuestionnaireResponse.QuestionnaireResponseItemComponent>,
   ) {
-    qItems.forEach {qit->
+    qItems.forEach { qit ->
       // process each inner item list
       deepFlat(qit.item, questionnaireResponse, targetQ, targetQR)
 
@@ -225,32 +244,37 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
       qit.item.clear()
 
       // add questionnaire and response pair for each linkid on same index
-      questionnaireResponse.find(qit.linkId)?.let {qrit->
+      questionnaireResponse.find(qit.linkId)?.let { qrit ->
         targetQ.add(qit)
         targetQR.add(qrit)
       }
     }
   }
 
+  // TODO change this when SDK bug for validation is fixed
+  // https://github.com/google/android-fhir/issues/912
   fun validQuestionnaireResponse(questionnaireResponse: QuestionnaireResponse): Boolean {
     // clone questionnaire and response for processing and changing structure
     val q = parser.parseResource(parser.encodeResourceToString(questionnaire)) as Questionnaire
-    val qr = parser.parseResource(parser.encodeResourceToString(questionnaireResponse)) as QuestionnaireResponse
+    val qr =
+      parser.parseResource(parser.encodeResourceToString(questionnaireResponse)) as
+        QuestionnaireResponse
 
-    // flatten and pair all responses temporarily to fix index mapping issue for questionnaire and questionnaire response
+    // flatten and pair all responses temporarily to fix index mapping issue for questionnaire and
+    // questionnaire response
     val qItems = mutableListOf<Questionnaire.QuestionnaireItemComponent>()
     val qrItems = mutableListOf<QuestionnaireResponse.QuestionnaireResponseItemComponent>()
 
     deepFlat(q.item, qr, qItems, qrItems)
 
-    return kotlin.runCatching {
-      QuestionnaireResponseValidator.validate(qItems, qrItems, this)
-        .values
-        .flatten()
-        .all { it.isValid }
-    }.onFailure {
-      Timber.e("Exception while validating response. Check if SDK issue with validation is fixed i.e. answer[0] at L56, $it")
-    }.getOrDefault(true)
+    return QuestionnaireResponseValidator.validate(
+      questionnaire.item,
+      questionnaireResponse.item,
+      this
+    )
+      .values
+      .flatten()
+      .all { it.isValid }
   }
 
   open fun handleQuestionnaireResponse(questionnaireResponse: QuestionnaireResponse) {
@@ -267,13 +291,21 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
     const val QUESTIONNAIRE_POPULATION_RESOURCES = "questionnaire-population-resources"
     const val QUESTIONNAIRE_FRAGMENT_TAG = "questionnaire-fragment-tag"
     const val QUESTIONNAIRE_ARG_PATIENT_KEY = "questionnaire_patient_item_id"
+    const val ADVERSE_EVENT_IMMUNIZATION_ITEM_KEY = "adverse_event_immunization_item_id"
     const val QUESTIONNAIRE_ARG_FORM = "questionnaire_form"
     const val FORM_CONFIGURATIONS = "form_configurations.json"
+    const val QUESTIONNAIRE_ARG_BARCODE_KEY = "patient-barcode"
+    const val WHO_IDENTIFIER_SYSTEM = "WHO-HCID"
 
-    fun requiredIntentArgs(clientIdentifier: String?, form: String) =
+    fun requiredIntentArgs(
+      clientIdentifier: String?,
+      form: String,
+      immunizationId: String? = null
+    ) =
       bundleOf(
         Pair(QUESTIONNAIRE_ARG_PATIENT_KEY, clientIdentifier),
         Pair(QUESTIONNAIRE_ARG_FORM, form),
+        Pair(ADVERSE_EVENT_IMMUNIZATION_ITEM_KEY, immunizationId)
       )
   }