+++ android/quest/src/test/java/org/smartregister/fhirecore/quest/ui/patient/register/PatientRegisterActivityTest.kt
@@ -18,7 +18,6 @@ package org.smartregister.fhirecore.quest.ui.patient.register
 
 import android.app.Activity
 import android.app.Application
-import android.content.Intent
 import android.view.View
 import android.widget.ImageButton
 import android.widget.TextView
@@ -29,13 +28,12 @@ import org.junit.Before
 import org.junit.BeforeClass
 import org.junit.Test
 import org.robolectric.Robolectric
-import org.robolectric.Shadows
 import org.robolectric.annotation.Config
 import org.robolectric.fakes.RoboMenuItem
 import org.robolectric.util.ReflectionHelpers
 import org.smartregister.fhircore.engine.configuration.view.loadRegisterViewConfiguration
 import org.smartregister.fhircore.engine.databinding.BaseRegisterActivityBinding
-import org.smartregister.fhircore.engine.ui.questionnaire.QuestionnaireActivity
+import org.smartregister.fhircore.engine.ui.register.model.RegisterItem
 import org.smartregister.fhircore.engine.ui.userprofile.UserProfileFragment
 import org.smartregister.fhircore.quest.R
 import org.smartregister.fhircore.quest.ui.patient.register.PatientRegisterActivity
@@ -108,24 +106,6 @@ class PatientRegisterActivityTest : ActivityRobolectricTest() {
   }
 
   @Test
-  fun testRegisterClientShouldStartFamilyQuestionnaireActivity() {
-    ReflectionHelpers.callInstanceMethod<PatientRegisterActivity>(
-      patientRegisterActivity,
-      "registerClient"
-    )
-
-    val expectedIntent = Intent(patientRegisterActivity, QuestionnaireActivity::class.java)
-    val actualIntent =
-      Shadows.shadowOf(ApplicationProvider.getApplicationContext<Application>()).nextStartedActivity
-
-    Assert.assertEquals(expectedIntent.component, actualIntent.component)
-    Assert.assertEquals(
-      "3435",
-      actualIntent.getStringExtra(QuestionnaireActivity.QUESTIONNAIRE_ARG_FORM)
-    )
-  }
-
-  @Test
   fun testSetupConfigurableViewsShouldUpdateViews() {
     val config =
       ApplicationProvider.getApplicationContext<Application>()
@@ -168,6 +148,21 @@ class PatientRegisterActivityTest : ActivityRobolectricTest() {
     Assert.assertEquals(2, fragments.size)
     Assert.assertTrue(fragments.containsKey(PatientRegisterFragment.TAG))
     Assert.assertTrue(fragments.containsKey(UserProfileFragment.TAG))
+  }
+
+  @Test
+  fun testRegistersListShouldReturnOnlyOneItemList() {
+    val list =
+      ReflectionHelpers.callInstanceMethod<List<RegisterItem>>(
+        patientRegisterActivity,
+        "registersList"
+      )
+    Assert.assertEquals(1, list.size)
+    with(list[0]) {
+      Assert.assertEquals(PatientRegisterFragment.TAG, uniqueTag)
+      Assert.assertEquals(patientRegisterActivity.getString(R.string.clients), title)
+      Assert.assertTrue(isSelected)
+    }
   }
 
   override fun getActivity(): Activity {